+++ NEWS.md
@@ -12,6 +12,7 @@
 - add `GramSchmidt()` for the Gram-Schmidt algorithm on columns of a matrix. The existing function `gsorth()` will be deprecated and then removed.
 - `gsorth()` has been deprecated.
 - fixed use of MASS::fractions in gaussianElimination
+- added `print_mat()` to print matrix expressions side-by-side
 
 # matlib 0.8.1